Attempt to simplify the situation regarding to the complex explicit template instantiation we had for windows (basically move everything to the headers).
