+++ OpenMEEG/include/assembleSourceMat.h
@@ -16,11 +16,6 @@
 
 #include <constants.h>
 
-#ifdef WIN32
-#include "dipole.h"
-#include "monopole.h"
-#endif
-
 namespace OpenMEEG {
 
     // For ADAPT_LHS change the 0 in Integrator below into 10
@@ -70,11 +65,6 @@ namespace OpenMEEG {
     Matrix SourceMatrix(const Geometry& geo,const Matrix& sources,const std::string& domain_name) {
         return SourceMatrix<Source>(geo,sources,Integrator(3,10,0.001),domain_name);
     }
-
-    #ifdef WIN32
-    template OPENMEEG_EXPORT Matrix SourceMatrix<Monopole>(const Geometry& geo,const Matrix& sources,const Integrator& integrator,const std::string& domain_name);
-    template OPENMEEG_EXPORT Matrix SourceMatrix<Dipole>(const Geometry& geo,const Matrix& sources,const Integrator& integrator,const std::string& domain_name);
-    #endif
 
     OPENMEEG_EXPORT Matrix EITSourceMat(const Geometry& geo,const Sensors& electrodes,const Integrator& integrator=Integrator(3,0,0.005));
     OPENMEEG_EXPORT Matrix DipSource2InternalPotMat(const Geometry& geo,const Matrix& dipoles,const Matrix& points,const std::string& domain_name="");
